Prebuilt client/server/python simulator
Added the built driver and example python HID simulators to the project.

Driver is currently self-signed - so secure boot settings may need to be modified on the install machine to accept self-signed drivers.

For real world deployment, we'll need to purchase a signing certificate from Verisign to properly sign the driver.